fix medico

+++ src/medico/medico.service.ts
@@ -47,6 +47,12 @@ export class MedicoService {
         BusinessError.NOT_FOUND,
       );
     }
+    if (medico.pacientes.length >= 1) {
+      throw new BusinessLogicException(
+        'No se puede eliminar un médico con pacientes asociados',
+        BusinessError.PRECONDITION_FAILED,
+      );
+    }
     await this.medicoRepository.delete(medico.id);
   }
 }
